requirements.txt
numpy==1.26.4
scipy==1.15.3
scikit-learn==1.4.1.post1
joblib==1.5.3
threadpoolctl==3.6.0

faiss-cpu==1.8.0

langchain==0.1.14
langchain-community==0.0.30
pypdf==4.1.0

google-genai==1.66.0